requirements.txt
# Core framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
PyYAML==6.0.1

# Kafka
confluent-kafka==2.4.0

# MongoDB
motor==3.4.0
pymongo==4.7.2

# Redis
redis==5.0.4

# Image processing
Pillow==10.3.0
numpy==1.26.4

# OCR
pytesseract==0.3.13

# HTTP client
httpx==0.27.0

# OpenTelemetry
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-httpx==0.45b0
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-exporter-otlp-proto-grpc==1.24.0

# Prometheus
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0

# Cloud SDKs (optional, install as needed)
boto3==1.34.88
google-cloud-storage==2.16.0
azure-storage-blob==12.19.1
azure-identity==1.16.0
